Refactored test and file names.
